humanized_diff(); only works on V1, so switched back to that.
This elegantly uses merge() with the fact diff()-V1 heavily relies
on sorting to print out the table interleaved with the changes.
Next: use this to experiment; I have this theory that every update
corresponds to an adjacent add/delete, and this will give intuition.

This commit also includes a bunch of scrap code which is work
towards generalized diffu() (using linear programming!)
NOT TO BE USED, but I also don't want to lose it right now,
so it's going into source control.