Some replication architecture planning.
I decided during this that 'view.py' was a bad name, and went with 'replicant.py' instead,
which turned out to be what I was calling the old, WAL-reading approach. So,
I put the WAL approach in an attic along with a note about what was tried and why it didn't work.