Switch to using the logging module.
Also, localize connections; instead of a global C,
have a C where its actually used.
The logging pointed out to me that I was closing C once I was done
commandeering the DB, but Changes needs it until its lifetime is
over--which is supposed to be never, ideally. I realized this is silly
and just reconnected to the DB as needed (which isn't so bad, really).