Prototype of CacheSet
CacheSet is an in-javascript database system that I'm hacking together because
none of the others I considered feature edge-triggered processing.

The other JS DB options are:
{Crossfilter,PourOver} -- which only provide *one* view of a dataset at a time, and I need *many*.
PouchDB -- which might actually serve, I haven't tried it
{WebSQL,IndexedDB,DOM Storage} -- which don't provide event triggers, so rerendering requires polling (ie getting the browser to recompute a slice);
  anyway, the killer feature of these is in-browser persistence which we don't even want for our application.

PourOver promised edge-triggering, but its views provide only "update" which it
triggers whenever anything *might* change the view, not whenever something *does*
change the view, and "collection-change" which triggers on writes to the parent.
Plus, its views are tied too tightly to each other; a view in PourOver consists
of boolean operations performed on the MatchSets that are cached on each Filter.
Since MatchSets are cached once per *filter*, in order to make multiple views
of the same dataset you need to make multiple sets of "filter" objects, unless
you desire coordinated views (which was the PourOver author's use case).
Coordinated views are useful, but trapping us to them is painful.

CacheSet is largely inspired by PourOver's MatchSets, but fills in exactly the features I need.
Maybe it can be rolled into PourOver some day, or a joint PourOver-Crossfilter-