Simplify HDRJCacheSet so that it *is* a CacheSet, and we don't have the awkwardness of needing to dig two layers deep to extract values.
To match this simplification, rename "change" to "rerender" and make the CacheSet fire it itself.

It would be nice if CacheSet itself was arrayish, so that you could just say set[0] and set.indexOf(e) and set.filter and pass it directly into D3 and whatnot. Oh well.

Also make copious comments about the available operators.