Debugging why unregister() doesn't get called when clients leave.
forkit.py demonstrates that socat does indeed shut down EXEC processes
when the socket dies (tested with both TCP and Unix Domain). So something about
replicate.py is holding it open idenfinitely. In fact, replicate.py can outlive socat.

I could rewrite replicate.py as a socket app itself, and in it select() on both the input socket and the output socket, but that's a lot of work and it might not even be the proper bug.